rag-ingestion/requirements.txt
deepeval==3.8.4
openai==2.18.0
tabulate==0.9.0
google-generativeai==0.8.6
pymupdf4llm==0.2.9
python-multipart==0.0.22
fastapi==0.128.7
uvicorn==0.40.0
structlog==25.5.0
psycopg[binary]==3.3.2
psycopg_pool==3.3.0
pydantic==2.12.5
torch==2.10.0
transformers==4.57.6
httpx==0.28.1
instructor==1.14.5
tqdm==4.67.3
rich==14.3.2
python-dotenv==1.2.1
supabase==2.28.0
pydantic-settings==2.12.0
langchain-core==1.2.10
langchain==1.2.10
langchain-groq==1.1.2
langchain-google-genai==4.2.0
tenacity==9.1.4
dspy-ai==3.1.3
langchain-anthropic==1.3.2
langchain-openai==1.1.8
langchain-community==0.4.1
scikit-learn==1.8.0
numpy==2.4.2
networkx==3.6.1
